Simple form with React-hook-form
